Not all pins support edge detection.
Not all pins on the Orange Pi Prime support edge detection, because
not all have an associated interrupt. This is presented to the user
in sysfs as a permissions denied error accessing the edge file.

Catch this error early and let the user know that the pin they've
selected cannot do edge detection.