Added Hardware PWM Control
Added hardware PWM using sysfs based on https://developer.toradex.com/knowledge-base/pwm-linux

It has been created it as a class in order to create better tracking of PWM parameters, this is important as PWM has more features than a standard GPIO output.

This was developed and tested on an OrangePi PC Plus and works :smile:  however was not tested on other boards. Likely though it will work on any H3 chip. The only thing to bare in mind is that you need to use PWM0 pin. On the OPi PC+ board this is the middle pin of the Debug TTL UART pins, located between the power and HDMI. The documentation mentions a PWM1 at pin A6 (physical pin 7) but I could not figure out how to access that. I believe the DTC will have to be changed to access it.

The commands that are used to create and control the PWM signal are:
p = GPIO.PWM(PWM_pin, frequency_Hz, Duty_Cycle_Percent) to create the object
p.start_pwm()      To begin the output
p.duty_cycle(Duty_Cycle_Percent)  To adjust the duty cycle
p.change_frequency(frequency_Hz) To change the PWM frequency
p.pwm_polarity() To adjust the polarity, ie. changing duty cycle from percentage on to percentage off or vice versa
p.stop_pwm() To stop the PWM by setting the duty cycle to 0
p.pwm_close() To completely remove the PWM sysfs object.

For an OrangePi H3 board such as the PC + use PWM_pin = 0. As the PWM setup is different to the GPIO setup I didn't integrate the pin map present in this library. Its not an inconvenience however as this will likely be the same for all OPi H3 boards, but if you want to check what number you need to put in use:
"ls -l /sys/class/pwm/" This will show you all the PWMchips are on your system. it will be something like pwmchip0.
"echo 0 > /sys/class/pwm/pwmchip0/export"  this will create the PWM object. if an error returns saying no such device try 1 and so on. They are generally low numbers less than 5.
"ls /sys/class/pwm/pwmchip0" this will list all of the controls as well as the PWM object that was created.  The object is should  be called PWM0 (or PWM-whatever number worked)

The number following the PWM object name is the pin number you use in GPIO.PWM(PWM_pin, frequency_Hz, Duty_Cycle_Percent). It probably wouldn't be too difficult to make a list of pins for the various boards and use that as the pin mapping but I haven't.

I tried to leave a lot of comments as I was going along so hopefully the changes should make sense.